+++ lightstep/sdk/metric/internal/viewstate/viewstate.go
@@ -410,9 +410,9 @@ func buildView[N number.Any, Traits number.Traits[N]](behavior singleBehavior) l
 }
 
 // newSyncView returns a compiled synchronous instrument.  If the view
-// calls for delta temporality, a stateless instrument is returned,
+// calls for delta temporality, a lowmemory instrument is returned,
 // otherwise for cumulative temporality a stateful instrument will be
-// used.  I.e., Delta->Stateless, Cumulative->Stateful.
+// used.  I.e., Delta->Lowmemory, Cumulative->Stateful.
 func newSyncView[
 	N number.Any,
 	Storage any,
@@ -435,7 +435,7 @@ func newSyncView[
 		instrumentBase: metric, //nolint:govet
 	}
 	if behavior.tempo == aggregation.DeltaTemporality {
-		return &statelessSyncInstrument[N, Storage, Methods]{
+		return &lowmemorySyncInstrument[N, Storage, Methods]{
 			compiledSyncBase: instrument, //nolint:govet
 		}
 	}
@@ -487,8 +487,8 @@ func compileSync[N number.Any, Traits number.Traits[N]](behavior singleBehavior)
 
 // newAsyncView returns a compiled asynchronous instrument.  If the
 // view calls for delta temporality, a stateful instrument is
-// returned, otherwise for cumulative temporality a stateless
-// instrument will be used.  I.e., Cumulative->Stateless,
+// returned, otherwise for cumulative temporality a lowmemory
+// instrument will be used.  I.e., Cumulative->Lowmemory,
 // Delta->Stateful.
 func newAsyncView[
 	N number.Any,
@@ -518,11 +518,11 @@ func newAsyncView[
 				compiledAsyncBase: instrument, //nolint:govet
 			}
 		}
-		// Gauges fall through to the stateless behavior
+		// Gauges fall through to the lowmemory behavior
 		// regardless of delta temporality.
 	}
 
-	return &statelessAsyncInstrument[N, Storage, Methods]{
+	return &lowmemoryAsyncInstrument[N, Storage, Methods]{
 		compiledAsyncBase: instrument, //nolint:govet
 	}
 }